+++ test/cases/option-preload-split-chunk/expected/assets/js/954.15b39930.js
@@ -1,0 +1,1 @@
+(self.webpackChunk=self.webpackChunk||[]).push([[954],{954:()=>{console.log(">> about")}}]);