+++ test/cases/option-preload-split-chunk/expected/assets/js/home.b9105436.js
@@ -1,0 +1,1 @@
+"use strict";(self.webpackChunk=self.webpackChunk||[]).push([[177],{},s=>{s.O(0,[87],(()=>(87,s(s.s=87)))),s.O()}]);