+++ test/cases/option-preload-split-chunk/expected/assets/js/sport.99073c9d.js
@@ -1,0 +1,1 @@
+"use strict";(self.webpackChunk=self.webpackChunk||[]).push([[385],{},s=>{s.O(0,[954],(()=>(954,s(s.s=954)))),s.O()}]);