+++ test/cases/option-preload-split-chunk/src/views/pages/home/home.js
@@ -1,0 +1,3 @@
+import lorem from '@test-fixtures/lorem';
+
+console.log('>> Home: ', lorem.getTitle());